LineChatBot/requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
Deprecated==1.3.1
Flask==3.1.2
frozenlist==1.8.0
future==1.0.0
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.71.0
httplib2==0.31.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
line-bot-sdk==3.20.0
MarkupSafe==3.0.3
multidict==6.7.0
oauthlib==3.3.1
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==2.0.0
yarl==1.22.0
